Updateed tests
